get category from image
