+++ server/controllers/images.controller.js
@@ -5,15 +5,12 @@ const fs = require('fs');
 const detectBook = require('./BookDetection/handleDetectionData');
 const fetchData = require('./BookDetection/bookDetection');
 
-const automl = require('@google-cloud/automl').v1beta1;
-const prediction = require('../../visiontest');
+const prediction = require('../services/prediction');
 
 const client = new vision.ImageAnnotatorClient({
   keyFilename: `${__dirname}/../../googleKey.json`,
 });
-const clientX = new automl.PredictionServiceClient({
-  keyFilename: `${__dirname}/../../googleKey.json`,
-});
+
 const createImageName = () => {
   const dir = path.join(__dirname, '../images/');
   const newName = Date.now();
@@ -27,15 +24,14 @@ const imageToBase64 = (file) => {
 
 let image = createImageName();
 
-const transformData = (values, imageInBase64) => {
-
+const transformData = (values, imageInBase64, predict) => {
   const collection = {
+    category: 'default',
     labels: [],
     text: [],
     imageName: '',
     imageInBase64: '',
     book: '',
-    category: 1,
   };
   const labelsArray = [];
   const textArray = [];
@@ -47,18 +43,19 @@ const transformData = (values, imageInBase64) => {
       return null;
     });
   });
+
   collection.text = textArray;
   collection.imageName = image.replace(/^\D+/g, '');
   collection.imageInBase64 = imageInBase64;
   collection.labels = labelsArray;
   collection.book = null;
-  collection.category = 2;
+  collection.category = predict;
   return collection;
 };
 
 const Capture = (res) => {
-
   image = createImageName();
+
   // to take picture from external web cam add name of device  as parameter to nodeWebcam.create({})
   // Use another device
   const anotherCam = nodeWebcam.create();
@@ -69,21 +66,22 @@ const Capture = (res) => {
     } else {
       try {
         const promises = [
-          prediction.predict(image),
           client.labelDetection(image),
           client.textDetection(image),
         ];
 
         const values = await Promise.all(promises);
+        // send image to prediction client, predictions are made based
+        // on the custom trained model.
+        const predict = await prediction(image);
 
+        // const val = await Promise.all(predict);
         const imageInBase64 = imageToBase64(image);
-        console.log(values);
-        const responseData = transformData(values, imageInBase64);
+
+        const responseData = transformData(values, imageInBase64, predict);
 
         const responseText = responseData.text;
         const bookData = detectBook.filter(responseText);
-
-        console.log(predict);
 
         if (bookData) {
           responseData.book = await fetchData(bookData);
@@ -110,7 +108,9 @@ const Delete = (res, imageName) => {
   res.sendStatus(200);
 };
 
-
 module.exports = {
-  Capture, Send, Delete, transformData,
+  Capture,
+  Send,
+  Delete,
+  transformData,
 };
